ingestion/requirements.txt
qdrant-client
sentence-transformers
requests
beautifulsoup4
pypdf
tqdm
python-dotenv